+++ NEWS.md
@@ -1,3 +1,10 @@
+v.0.1.4
+=======
+* new plots for PLS model: `plotxloadings()`, `plotxyloadings()` and `plotweights()`
+* small improvements for regression coefficients plot
+* methods `scatter()` and `plot()` now have a new parameter `'Groupby'` for easy color grouping
+* method `catter()` has a new parameter `'ShowContour'` for drawing contour for a cluster of points
+
 v.0.1.3
 ========
 * fixed bug in `predict()` method for regression, which did not work if y references have not been provided